+++ CHANGELOG.md
@@ -10,6 +10,16 @@
 * restrict browsing until day
 
 
+# 0.7.12 (2026-02-18)
+
+- changed scores.clj - stop peep service.
+- updated library
+
+| :file    | :name       | :current | :latest |
+|----------|-------------|----------|---------|
+| deps.edn | babashka/fs | 0.5.30   | 0.5.31  |
+
+
 # 0.7.11 (2026-02-04)
 
 - ruff-0.15.0